Update and rename 03_grid_bot.py to grid_bot.py
